Test organization and fixup doc strings
